+++ dags/process_zip_dag.py
@@ -22,14 +22,14 @@ from lxml.etree import ElementTree
 from requests import HTTPError
 from wand.image import Image
 
-from aws import get_s3_client, list_bucket_keys_iter
-from libero.xml import jats, libero, XLINK_HREF, XLINK_MAP, XML_BASE
-from libero.xml.xpaths import XLINK_HREF_CONTAINS_TIF, XLINK_HREF_STARTS_WITH_WWW
-from task_helpers import (
+from libero.aws import get_s3_client, list_bucket_keys_iter
+from libero.task_helpers import (
     get_file_name_passed_to_dag_run_conf_file,
     get_previous_task_name,
     get_return_value_from_previous_task
 )
+from libero.xml import jats, libero, XLINK_HREF, XLINK_MAP, XML_BASE
+from libero.xml.xpaths import XLINK_HREF_CONTAINS_TIF, XLINK_HREF_STARTS_WITH_WWW
 
 
 # settings
